test: split visual method listing regression
